Crate amfi

amfi: library to fetch latest NAV data from AMFI

It aims to extract as much information from AMFI latest nav public data as possible.

This library can also parse data mirrors and local file copies. See nav_from_url and nav_from_file.

Basic Usage

let navs = amfi::daily_nav();
for item in items {
    match item {
        Err(error) => warn!("{}", error),
        Ok(ref record) => println!("{:>10} {} {}", record.nav, record.date, record.name),
    }
}

Cargo features

Enable serde feature for serialization/deserialization support.

Structs

NavRecord
Net Asset Value Record

NavRecordIterator
Iterator over NavRecord

Enums

Error
Error type

FundMaturity
Open/Closed Funds

FundPlan
Fund Plans are identified on best effort basis. By default plans are Regular.

Functions

daily_nav
Parses NAV data from AMFI portal

nav_from_file
Parses NAV data from local file

nav_from_url
Parses NAV data from provided url